--- a/pytest.ini
+++ b/pytest.ini
@@ -22,6 +22,9 @@ python_files = test_*.py
 python_classes = Test*
 python_functions = test_*
 
+# Ignore property tests when hypothesis not installed
+norecursedirs = tests/property
+
 # ============================================================================
 # DEFAULT BEHAVIOR: FAST TDD FEEDBACK LOOP
 # ============================================================================
@@ -44,10 +47,6 @@ addopts =
     --strict-markers
     --tb=short
     --color=yes
-    --reruns 3
-    --reruns-delay 1
-    --json-report
-    --json-report-file=test-results/report.json
 
 # ============================================================================
 # PERFORMANCE OPTIMIZATION SETTINGS