… .snapshots/AUTONOMOUS_EXECUTION_SUMMARY.md
… .snapshots/MERGE_STATUS.md
… .snapshots/NEXT_10X_OPPORTUNITIES.md
… .snapshots/PHASE1_QUICK_WINS_COMPLETE.md
… logs/constitutional_telemetry/events_20251008.jsonl
… messages.db
… shared/model_policy.py
… test-results/report.json
… tests/property/test_critical_properties.py
… tests/property/test_property_based.py
… tests/test_agent_context_caching.py
… tests/test_continuous_audit.py
… tests/test_firestore_learning_persistence.py
… tests/test_model_policy_enhanced.py
… tests/test_model_routing.py